src/backend/requirements.txt
aiofiles==23.1.0
anyio==3.6.2
bcrypt==4.0.1
certifi==2023.5.7
click==8.1.3
h11==0.14.0
html5tagger==1.3.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
nodeenv==1.8.0
prisma==0.8.2
pydantic==1.10.8
PyJWT==2.7.0
python-dotenv==1.0.0
sanic==23.3.0
sanic-routing==22.8.0
sniffio==1.3.0
tomlkit==0.11.8
tracerite==1.1.0
typing_extensions==4.6.1
ujson==5.7.0
uvloop==0.17.0
websockets==11.0.3
